Simplify star data and chart performance
Moved rendering of stars to their own component. Since the stars will
never need to update, Stars component will never update via
`shouldComponentUpdate`.

Because of a naming conflict, I had to change `Stars` to `darData`. I
refactored out the object there because there used to be functions in
that object. Now that it's just data I switched it up to an array.